# Add `@stub` annotation to mark stubs