--- a/javalib/src/main/scala/java/lang/ref/ReferenceQueue.scala
+++ b/javalib/src/main/scala/java/lang/ref/ReferenceQueue.scala
@@ -1,5 +1,6 @@
 package java.lang.ref
 
 class ReferenceQueue[T >: Null <: AnyRef] {
+  @scala.scalanative.native.stub
   def poll(): java.lang.ref.Reference[_] = null
 }