--- a/javalib/src/main/scala/java/lang/reflect/Constructor.scala
+++ b/javalib/src/main/scala/java/lang/reflect/Constructor.scala
@@ -1,8 +1,12 @@
 package java.lang
 package reflect
 
+import scala.scalanative.native.stub
+
 class Constructor[T] extends Executable {
+  @stub
   def getParameterTypes(): scala.Array[Object] = ???
+  @stub
   def newInstance(
       args: scala.scalanative.runtime.ObjectArray): java.lang.Object = ???
 }