--- a/javalib/src/main/scala/java/lang/reflect/Field.scala
+++ b/javalib/src/main/scala/java/lang/reflect/Field.scala
@@ -1,12 +1,7 @@
 package java.lang.reflect
 
-import scala.scalanative.native.stub
-
 class Field {
-  @stub
   def get(obj: Object): Object = ???
-  @stub
-  def getName(): String = ???
-  @stub
-  def getType(): Class[_] = ???
+  def getName(): String        = ???
+  def getType(): Class[_]      = ???
 }