--- a/javalib/src/main/scala/java/lang/reflect/Method.scala
+++ b/javalib/src/main/scala/java/lang/reflect/Method.scala
@@ -1,10 +1,17 @@
 package java.lang.reflect
 
+import scala.scalanative.native.stub
+
 class Method {
-  def getDeclaringClass(): java.lang.Class[_]  = ???
-  def getName(): java.lang.String              = ???
+  @stub
+  def getDeclaringClass(): java.lang.Class[_] = ???
+  @stub
+  def getName(): java.lang.String = ???
+  @stub
   def getParameterTypes(): scala.Array[Object] = ???
-  def getReturnType(): java.lang.Class[_]      = ???
+  @stub
+  def getReturnType(): java.lang.Class[_] = ???
+  @stub
   def invoke(obj: java.lang.Object,
              args: scala.Array[Object]): java.lang.Object = ???
 }