--- a/javalib/src/main/scala/java/net/URI.scala
+++ b/javalib/src/main/scala/java/net/URI.scala
@@ -1163,6 +1163,6 @@ final class URI private () extends Comparable[URI] with Serializable {
     convertHexToLowerCase(result.toString)
   }
 
-  @scala.scalanative.native.stub
+  @scalanative.native.stub
   def toURL(): java.net.URL = ???
 }