--- a/javalib/src/main/scala/java/net/URL.scala
+++ b/javalib/src/main/scala/java/net/URL.scala
@@ -1,6 +1,6 @@
 package java.net
 
-import scala.scalanative.native.stub
+import scalanative.native.stub
 
 class URL(from: String) {
   @stub