--- a/javalib/src/main/scala/java/net/URLConnection.scala
+++ b/javalib/src/main/scala/java/net/URLConnection.scala
@@ -1,5 +1,6 @@
 package java.net
 
 class URLConnection {
+  @scala.scalanative.native.stub
   def getLastModified(): scala.Long = ???
 }