--- a/javalib/src/main/scala/java/nio/channels/Channels.scala
+++ b/javalib/src/main/scala/java/nio/channels/Channels.scala
@@ -2,7 +2,7 @@ package java.nio
 package channels
 
 object Channels {
-  @scala.scalanative.native.stub
+  @scalanative.native.stub
   def newChannel(
       is: java.io.InputStream): java.nio.channels.ReadableByteChannel = ???
 }