--- a/nir/src/main/scala/scala/scalanative/nir/serialization/BinarySerializer.scala
+++ b/nir/src/main/scala/scala/scalanative/nir/serialization/BinarySerializer.scala
@@ -70,8 +70,7 @@ final class BinarySerializer(buffer: ByteBuffer) {
     case Attr.NoInline     => putInt(T.NoInlineAttr)
     case Attr.AlwaysInline => putInt(T.AlwaysInlineAttr)
 
-    case Attr.Dyn => putInt(T.DynAttr)
-
+    case Attr.Dyn  => putInt(T.DynAttr)
     case Attr.Stub => putInt(T.StubAttr)
 
     case Attr.Align(_) =>