--- a/nscplugin/src/main/scala/scala/scalanative/nscplugin/NirGenStat.scala
+++ b/nscplugin/src/main/scala/scala/scalanative/nscplugin/NirGenStat.scala
@@ -132,6 +132,8 @@ trait NirGenStat { self: NirGenPhase =>
         case ann if ann.symbol == LinkClass =>
           val Apply(_, Seq(Literal(Constant(name: String)))) = ann.tree
           Attr.Link(name)
+        case ann if ann.symbol == StubClass =>
+          Attr.Stub
       }
       val pure = if (PureModules.contains(sym)) Seq(Attr.Pure) else Seq()