--- a/sbt-scala-native/src/main/scala/scala/scalanative/sbtplugin/ScalaNativePlugin.scala
+++ b/sbt-scala-native/src/main/scala/scala/scalanative/sbtplugin/ScalaNativePlugin.scala
@@ -31,6 +31,9 @@ object ScalaNativePlugin extends AutoPlugin {
       taskKey[Seq[String]](
         "Additional options that are passed to clang during linking.")
 
+    val nativeLinkStubs =
+      settingKey[Boolean]("Whether to link `@stub` methods, or ignore them.")
+
     val nativeLink =
       taskKey[File]("Generates native binary without running it.")