--- a/scripted-tests/run/link-stubs/Main.scala
+++ b/scripted-tests/run/link-stubs/Main.scala
@@ -1,0 +1,9 @@
+object Main {
+  def main(args: Array[String]): Unit =
+    Foo.bar
+}
+
+object Foo {
+  @scala.scalanative.native.stub
+  def bar = ???
+}