--- a/scripted-tests/run/link-stubs/Main.scala
+++ b/scripted-tests/run/link-stubs/Main.scala
@@ -4,6 +4,6 @@ object Main {
 }
 
 object Foo {
-  @scala.scalanative.native.stub
+  @scalanative.native.stub
   def bar = ???
 }