--- a/scripted-tests/run/link-stubs/build.sbt
+++ b/scripted-tests/run/link-stubs/build.sbt
@@ -1,0 +1,3 @@
+enablePlugins(ScalaNativePlugin)
+
+scalaVersion := "2.11.11"