--- a/scripted-tests/run/link-stubs/test
+++ b/scripted-tests/run/link-stubs/test
@@ -1,0 +1,3 @@
+-> nativeLink
+> set nativeLinkStubs := true
+> nativeLink