--- a/tools/src/test/scala/scala/scalanative/nir/AttrParserTest.scala
+++ b/tools/src/test/scala/scala/scalanative/nir/AttrParserTest.scala
@@ -16,6 +16,7 @@ class AttrParserTest extends FlatSpec with Matchers {
       NoInline,
       AlwaysInline,
       Dyn,
+      Stub,
       Align(1024),
       Pure,
       Extern,