… docs/contrib/nir.rst
… docs/user/sbt.rst