By default, method marked with `@stub` will not link.

Fixes #971.