@densh 2017-08-30
One more thing: needs an update of `sbt.rst` to mention a new setting.