@Duhemm 2017-08-31
@densh I've added all the `@stub`. `java.util.Formatter` depends on many things that were just stubbed out in `Calendar` and `TimeZone`, so I had to comment that code in `Formatter` and open issues to track it.

Where do the dependencies on `Class.getInterfaces` and `Class.getSuperclass` come from?

Regarding documenting the option, it looks `nativeConfig` is not mentioned at all (it is in `ScalaNativePluginInternal`). Should I introduce a new key, say `scalaNativeLinkStubs` in the sbt plugin?