@densh 2017-08-31
We do need to expose an sbt setting. `nativeLinkStubs` sounds like a good fit for this.