@Duhemm 2017-08-31
Before:
```
> tests/nativetest:nativeLink
[info] Linking (3700 ms)
[info] Discovered 4137 classes and 23843 methods
[info] Optimizing (4420 ms)
[info] Generating intermediate code (2683 ms)
[info] Produced 66 files
[info] Compiling to native code (1319 ms)
[info] Linking native code (327 ms)
[success] Total time: 15 s, completed Aug 31, 2017 2:05:30 PM
> tests/nativetest:nativeLink
[info] Linking (3423 ms)
[info] Discovered 4137 classes and 23843 methods
[info] Optimizing (3241 ms)
[info] Generating intermediate code (2074 ms)
[info] Produced 66 files
[info] Compiling to native code (1390 ms)
[info] Linking native code (338 ms)
[success] Total time: 11 s, completed Aug 31, 2017 2:05:43 PM
> tests/nativetest:nativeLink
[info] Linking (3632 ms)
[info] Discovered 4137 classes and 23843 methods
[info] Optimizing (3547 ms)
[info] Generating intermediate code (1716 ms)
[info] Produced 66 files
[info] Compiling to native code (1267 ms)
[info] Linking native code (340 ms)
[success] Total time: 11 s, completed Aug 31, 2017 2:05:57 PM
> tests/nativetest:nativeLink
[info] Linking (3248 ms)
[info] Discovered 4137 classes and 23843 methods
[info] Optimizing (3445 ms)
[info] Generating intermediate code (1552 ms)
[info] Produced 66 files
[info] Compiling to native code (1215 ms)
[info] Linking native code (341 ms)
[success] Total time: 14 s, completed Aug 31, 2017 2:06:12 PM
> tests/nativetest:nativeLink
[info] Linking (2999 ms)
[info] Discovered 4137 classes and 23843 methods
[info] Optimizing (3549 ms)
[info] Generating intermediate code (1524 ms)
[info] Produced 66 files
[info] Compiling to native code (1356 ms)
[info] Linking native code (350 ms)
[success] Total time: 10 s, completed Aug 31, 2017 2:06:24 PM
```

After:
```
> tests/nativetest:nativeLink
[info] Linking (3881 ms)
[info] Discovered 4038 classes and 23253 methods
[info] Optimizing (4170 ms)
[info] Generating intermediate code (2568 ms)
[info] Produced 66 files
[info] Compiling to native code (1116 ms)
[info] Linking native code (312 ms)
[success] Total time: 14 s, completed Aug 31, 2017 1:55:20 PM
> tests/nativetest:nativeLink
[info] Linking (3310 ms)
[info] Discovered 4038 classes and 23253 methods
[info] Optimizing (3053 ms)
[info] Generating intermediate code (1766 ms)
[info] Produced 66 files
[info] Compiling to native code (1153 ms)
[info] Linking native code (325 ms)
[success] Total time: 10 s, completed Aug 31, 2017 1:55:32 PM
> tests/nativetest:nativeLink
[info] Linking (3080 ms)
[info] Discovered 4038 classes and 23253 methods
[info] Optimizing (3008 ms)
[info] Generating intermediate code (1360 ms)
[info] Produced 66 files
[info] Compiling to native code (1162 ms)
[info] Linking native code (320 ms)
[success] Total time: 9 s, completed Aug 31, 2017 1:55:44 PM
> tests/nativetest:nativeLink
[info] Linking (3342 ms)
[info] Discovered 4038 classes and 23253 methods
[info] Optimizing (2908 ms)
[info] Generating intermediate code (1292 ms)
[info] Produced 66 files
[info] Compiling to native code (1142 ms)
[info] Linking native code (320 ms)
[success] Total time: 12 s, completed Aug 31, 2017 1:56:02 PM
> tests/nativetest:nativeLink
[info] Linking (2644 ms)
[info] Discovered 4038 classes and 23253 methods
[info] Optimizing (2920 ms)
[info] Generating intermediate code (1247 ms)
[info] Produced 66 files
[info] Compiling to native code (1100 ms)
[info] Linking native code (346 ms)
[success] Total time: 9 s, completed Aug 31, 2017 1:56:12 PM
```