@Duhemm 2017-09-01
@densh I've added the timings, added the `@stub` annotation for everything that has been added post 0.3.2 and added a new setting key, `nativeLinkStub`.